requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.0
openai>=1.0.0
gspread>=6.0.0
google-auth>=2.20.0
requests>=2.31.0
python-dotenv>=1.0.0
